Renaming files and removed testing
