+++ ariac_gui/ariac_gui/trial_generator.py
